chore: Organize src/, add bin/
