+++ .macros.vim
@@ -1,3 +1,3 @@
-nnoremap @1 :tabnew<cr>:term python3 src/Server.py<cr>
-nnoremap @2 :tabnew<cr>:term python3 src/Client.py<cr>
-nnoremap @3 :tabnew<cr>:term python3 CSChat.py<cr>
+nnoremap @1 :tabnew<cr>:term bin/server.sh<cr>
+nnoremap @2 :tabnew<cr>:term bin/client.sh<cr>
+nnoremap @3 :tabnew<cr>:term python3 -m src.client.CSChat<cr>
