+++ bin/server.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+python3 -m src.server.Server
